Moved step functions out of MessageSendingDemo into their own file StepFunctions for cleaner access from SimulationLoop.
